Separate chat context from run supervision
